Updated JPMML-Converter dependency

+++ src/main/java/org/jpmml/sparkml/model/TreeModelUtil.java
@@ -40,6 +40,7 @@ import org.dmg.pmml.SimplePredicate;
 import org.dmg.pmml.True;
 import org.dmg.pmml.Visitor;
 import org.dmg.pmml.tree.BranchNode;
+import org.dmg.pmml.tree.ClassifierNode;
 import org.dmg.pmml.tree.LeafNode;
 import org.dmg.pmml.tree.Node;
 import org.dmg.pmml.tree.TreeModel;
@@ -54,7 +55,6 @@ import org.jpmml.converter.ModelUtil;
 import org.jpmml.converter.PredicateManager;
 import org.jpmml.converter.Schema;
 import org.jpmml.converter.ValueUtil;
-import org.jpmml.converter.tree.ClassifierNode;
 import org.jpmml.sparkml.ModelConverter;
 import org.jpmml.sparkml.visitors.TreeModelCompactor;
 
@@ -135,7 +135,7 @@ public class TreeModelUtil {
 
 					ImpurityCalculator impurityCalculator = leafNode.impurityStats();
 
-					node.setRecordCount((double)impurityCalculator.count());
+					node.setRecordCount(impurityCalculator.count());
 
 					List<ScoreDistribution> scoreDistributions = node.getScoreDistributions();
